ai.qwen-image-local/requirements_api.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
torch>=2.0.0
diffusers>=0.30.0
transformers>=4.51.3
pillow>=10.0.0
accelerate>=0.24.0
